% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLprcp-package.R
\docType{package}
\name{MLprcp-package}
\alias{MLprcp}
\alias{MLprcp-package}
\title{MLprcp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
