% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_randomForest.R
\name{randomForest_kcv}
\alias{randomForest_kcv}
\title{k-fold randomForest}
\usage{
randomForest_kcv(XX, YY, kfold = 6, seed = 1, ind_all = NULL, ntree = 500, ...)
}
\arguments{
\item{...}{others to \code{\link[randomForest:randomForest]{randomForest::randomForest()}}}
}
\description{
k-fold randomForest
}
\examples{
randomForest_kcv(XX, YY, kfold = 6, seed = 1)
}
